Build fix: help OW 1.4 to apply template.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@37805 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/datstrm.cpp
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// Name:        datstrm.cpp
+// Name:        src/common/datstrm.cpp
 // Purpose:     Data stream classes
 // Author:      Guilhem Lavaux
 // Modified by: Mickael Gilabert
@@ -296,7 +296,7 @@ void wxDataInputStream::ReadLL(wxLongLong *buffer, size_t size)
 wxLongLong wxDataInputStream::ReadLL(void)
 {
     wxLongLong ll;
-    DoReadLL(&ll, 1, m_input, m_be_order);
+    DoReadLL(&ll, (size_t)1, m_input, m_be_order);
     return ll;
 }
 #endif // wxUSE_LONGLONG
@@ -724,4 +724,3 @@ wxDataOutputStream& wxDataOutputStream::operator<<(float f)
 
 #endif
   // wxUSE_STREAMS
-
